% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{cytoscapeout}
\alias{cytoscapeout}
\title{getverboseplot.all
module membership.}
\usage{
cytoscapeout(
  datExpr = exp.ds$table2,
  power = exp.ds$power,
  module = exp.ds$mdl,
  moduleColors = exp.ds$moduleColors,
  threshold = exp.ds$threshold
)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{power}{power value.}

\item{module}{which module you choose to observation.}

\item{moduleColors}{param generated by step getnetwork.}

\item{threshold}{threshold of weight}
}
\value{
A list contains gene module membership and Pvalue.
}
\description{
getverboseplot.all
module membership.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
