% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getdatExpr}
\alias{getdatExpr}
\title{Data cleaning of expression matrix.
Follow the rules of https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html.}
\usage{
getdatExpr(rawdata, RcCutoff, samplePerc, datatype, method)
}
\arguments{
\item{rawdata}{Input expression data, such as FPKM, CPM, TPM or read count, expected count, peak area or protein abundance.}

\item{RcCutoff}{Cutoff of rawdata, if readcount or FPKM less than this "RcCutoff" in x samples, it should be treat as the noise.}

\item{samplePerc}{If readcount or FPKM less than RcCutoff "insamplePerc"  samples, it should be treat as the noise.}

\item{datatype}{Data type of your input rawdata. "count", "expected count". "normalized count", "peak area" or "protein abundance"}

\item{method}{Conversion method of input raw data. "vst" for count and expected count, "raw" and "logarithm" for other datatype.}
}
\value{
dx a matrix like datExpr
}
\description{
Data cleaning of expression matrix.
Follow the rules of https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html.
}
\references{
https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html
}
\author{

}
