% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getdatExpr2}
\alias{getdatExpr2}
\title{Data cleaning of expression matrix step2.
Follow the rules of https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html.}
\usage{
getdatExpr2(datExpr, GeneNumCut, cutmethod)
}
\arguments{
\item{datExpr}{Expression data matrix generate from 1st step.}

\item{GeneNumCut}{How many transcripts you want to retain for WGCNA.}

\item{cutmethod}{The datExpr will ranked by MAD or Var. which method you want.}
}
\value{
dx a matrix like datExpr
}
\description{
Data cleaning of expression matrix step2.
Follow the rules of https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html.
}
\references{
https://www.jianshu.com/p/e9cc3f43441d
}
\author{

}
