% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getKME}
\alias{getKME}
\title{Get kME.
kME values export.}
\usage{
getKME(datExpr, moduleColors, MEs_col)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{moduleColors}{module to color.}

\item{MEs_col}{module eigengene.}
}
\value{
A list contains multiple parameter for next step.
}
\description{
Get kME.
kME values export.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
