% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getMM}
\alias{getMM}
\title{Get MM.
module membership.}
\usage{
getMM(datExpr, MEs_col, nSamples, corType)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{MEs_col}{module eigengene.}

\item{nSamples}{module to color.}

\item{corType}{correlation types. such as "pearson"}
}
\value{
A list contains gene module membership and Pvalue.
}
\description{
Get MM.
module membership.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
