% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getMt}
\alias{getMt}
\title{Module-trait prepare.
Generate the variables needed by the module-trait relationships.}
\usage{
getMt(phenotype, nSamples, moduleColors, datExpr)
}
\arguments{
\item{phenotype}{Phenotype matrix, with sample in row and traits in column.}

\item{moduleColors}{module to color.}

\item{datExpr}{datExpr produced by 2nd step.}

\item{MEs_col}{module eigengene.}
}
\value{
A list contains multiple parameter for next step.
}
\description{
Module-trait prepare.
Generate the variables needed by the module-trait relationships.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
