% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getnetwork}
\alias{getnetwork}
\title{Scale-free network generation.
By WGCNA::blockwiseModules}
\usage{
getnetwork(datExpr, power, minModuleSize, mergeCutHeight, maxBlocksize)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{power}{power you select after scale-free test.}

\item{minModuleSize}{minimal size of module.}

\item{mergeCutHeight}{Dynamic cutting height, clusters less than this value will be merged}

\item{maxBlocksize}{number of select genes}
}
\value{
A list contains multiple parameter for next step.
}
\description{
Scale-free network generation.
By WGCNA::blockwiseModules
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
