% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getpower}
\alias{getpower}
\title{SFT detection.
Calculate the appropriate power for scale-free network construction.}
\usage{
getpower(datExpr, rscut)
}
\arguments{
\item{datExpr}{Expression data matrix generate from 2nd step.}

\item{rscut}{cutoff of R2.}
}
\value{
A list contains power test detials, power and two plots which demonstrated the sft and mean connectivity.
}
\description{
SFT detection.
Calculate the appropriate power for scale-free network construction.
}
\references{
https://www.jianshu.com/p/e9cc3f43441d
}
\author{

}
