% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getsampleTree}
\alias{getsampleTree}
\title{Plot the sample tree.
The function of the number of sample clusters is to remove outliers.}
\usage{
getsampleTree(datExpr, layout = "circular")
}
\arguments{
\item{datExpr}{Expression data matrix generate from 2nd step.}

\item{layout}{Layout of your sample tree cluster.}
}
\value{
A list contains Tree file, and tree plot.
}
\description{
Plot the sample tree.
The function of the number of sample clusters is to remove outliers.
}
\references{
https://www.jianshu.com/p/e9cc3f43441d
}
\author{

}
