% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getverboseplot.all}
\alias{getverboseplot.all}
\title{getverboseplot.all
module membership.}
\usage{
getverboseplot.all(
  datExpr,
  module,
  pheno,
  traitData,
  moduleColors,
  geneModuleMembership,
  MEs,
  nSamples
)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{module}{which module you choose to observation.}

\item{pheno}{which traits you choose to observation.}

\item{traitData}{trait matrix.}

\item{moduleColors}{param generated by step getnetwork.}

\item{geneModuleMembership}{param generated by getMM}

\item{MEs}{param generated by getnetwork}

\item{nSamples}{how many samples.}
}
\value{
A list contains gene module membership and Pvalue.
}
\description{
getverboseplot.all
module membership.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
