% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{ggscaleFreePlot}
\alias{ggscaleFreePlot}
\title{ggplot2 version of scaleFreePlot.
Transformed from the WGCNA::scaleFreePlot function.}
\usage{
ggscaleFreePlot(
  connectivity,
  nBreaks = 10,
  truncated = FALSE,
  removeFirst = FALSE,
  main = "",
  ...
)
}
\arguments{
\item{connectivity}{See WGCNA::scaleFreePlot.}

\item{nBreaks}{See WGCNA::scaleFreePlot.}

\item{truncated}{See WGCNA::scaleFreePlot}
}
\value{
A list contains Tree file, and tree plot.
}
\description{
ggplot2 version of scaleFreePlot.
Transformed from the WGCNA::scaleFreePlot function.
}
\references{
https://rdrr.io/cran/WGCNA/man/scaleFreePlot.html
}
\author{

}
