% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{hubgenes}
\alias{hubgenes}
\title{select hubgene
select hubgenes with GS and MM}
\usage{
hubgenes(datExpr, mdl, power, trt, KME, GS.cut, kME.cut, datTrait, g2m)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{mdl}{which module you choose to observation.}

\item{power}{power you selected.}

\item{trt}{trait you selected.}

\item{KME}{param generated by.}

\item{GS.cut}{cutoff Gene significance.}

\item{kME.cut}{cutoff kME.}

\item{datTrait}{trait matrix.}

\item{g2m}{gene2module.}
}
\value{
A list of hubgenes.
}
\description{
select hubgene
select hubgenes with GS and MM
}
\author{

}
