% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{iterative_out}
\alias{iterative_out}
\title{remove outlier
module membership.}
\usage{
iterative_out(g2m, rawMat, tbl, method, KME_cutoff)
}
\arguments{
\item{g2m}{dataframe label each gene with module tag.}

\item{rawMat}{dataframe original expression matrix}

\item{tbl}{dataframe KME table}

\item{method}{numeric method1 or method2.}

\item{KME_cutoff}{numeric kme cutoff.}
}
\value{
A list with retain and removed gene expression data.
}
\description{
remove outlier
module membership.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
