% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{MMvsGSall}
\alias{MMvsGSall}
\title{GS vs MM.all
module membership.}
\usage{
MMvsGSall(
  which.trait = exp.ds$st,
  traitData = exp.ds$phen,
  datExpr = exp.ds$table2,
  moduleColors = exp.ds$moduleColors,
  geneModuleMembership = exp.ds$MM,
  MEs = exp.ds$MEs_col,
  nSamples = exp.ds$nSample
)
}
\arguments{
\item{which.trait}{which trait you choose to observation.}

\item{traitData}{trait matrix.}

\item{datExpr}{datExpr produced by 2nd step.}

\item{moduleColors}{param generated by step getnetwork.}

\item{geneModuleMembership}{param generated by getMM}

\item{MEs}{param generated by getnetwork}

\item{nSamples}{how many samples.}
}
\value{
A list contains gene module membership and Pvalue.
}
\description{
GS vs MM.all
module membership.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
