% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{moduleheatmap}
\alias{moduleheatmap}
\title{Module heatmap.
Gene in module heatmap.}
\usage{
moduleheatmap(datExpr, MEs, which.module, moduleColors)
}
\arguments{
\item{datExpr}{datExpr produced by 2nd step.}

\item{MEs}{param generated by getnetwork}

\item{which.module}{which module you choose to observation.}

\item{moduleColors}{param generated by step getnetwork.}
}
\value{
A heatmap of selected module.
}
\description{
Module heatmap.
Gene in module heatmap.
}
\author{

}
