% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{mv_outlier}
\alias{mv_outlier}
\title{remove outlier
module membership.}
\usage{
mv_outlier(x, y)
}
\arguments{
\item{x}{dataframe original expression matrix}

\item{y}{character outlier samples.}
}
\value{
A dataframe.
}
\description{
remove outlier
module membership.
}
\references{
https://www.jianshu.com/p/f0409a045dab
}
\author{

}
