% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{powertest}
\alias{powertest}
\title{Power test.
Test scale-free of corresponding power}
\usage{
powertest(power.test, datExpr, nGenes)
}
\arguments{
\item{power.test}{Power you want test.}

\item{datExpr}{datExpr generate in second step.}

\item{nGenes}{Select gene number}
}
\value{
Plot of scale-free test result
}
\description{
Power test.
Test scale-free of corresponding power
}
\references{
https://www.jianshu.com/p/25905a905086 jiashu Author: sxzt
}
\author{

}
