% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{getdatExpr}
\alias{getdatExpr}
\title{Data cleaning of expression matrix.
Follow the rules of https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html.}
\usage{
getdatExpr(rawdata, RcCutoff, samplePerc, datatype, method)
}
\arguments{
\item{rawdata}{Input expression data, such as FPKM, CPM, TPM or read count.}

\item{RcCutoff}{Cutoff of rawdata, if readcount or FPKM less than this "RcCutoff" in x samples, it should be treat as the noise.}

\item{samplePerc}{If readcount or FPKM less than RcCutoff "insamplePerc"  samples, it should be treat as the noise.}

\item{datatype}{Data type of your input rawdata. readcount or FPKM. CPM,TPM,RPM and RPKM = "FPKM"}

\item{method}{Conversion method of input raw data. readcount: "varianceStabilizingTransformation" or "lgcpm", FPKM: "rawFPKM" or "lgFPKM"}
}
\value{
dx a matrix like datExpr
}
\description{
Data cleaning of expression matrix.
Follow the rules of https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html.
}
\references{
https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html
}
\author{

}
